+++ public/js/components/sorter.js
@@ -6,10 +6,10 @@ let compiledTemplate = _.template(template);
 import Component from '../component.js';
 
 export default class Sorter extends Component {
-  constructor({ element, list: optionsList }) {
+  constructor({ element, options }) {
     super(element);
 
-    this._list = optionsList;
+    this._options = options;
 
     this._render();
 
@@ -22,7 +22,7 @@ export default class Sorter extends Component {
 
   _render() {
     this._element.innerHTML = compiledTemplate({
-      list: this._list,
+      list: this._options,
     });
   }
 }
